+++ Functions/Settings/getSettings.js
@@ -1,0 +1,7 @@
+module.exports = getSettings
+
+const fs = require('fs')
+function getSettings() {
+    let settings = fs.readFileSync(`./esi.json`,'Utf8')
+    return settings;
+}