+++ Functions/test.js
@@ -1,5 +1,0 @@
-module.exports = Test 
-
-function Test(a, b, c) {
-    console.log(`A: ${a}, B: ${b}, C: ${c}`)
-}